backend/requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-dotenv==0.19.0
beautifulsoup4==4.9.3
requests==2.26.0
# pyenchant==3.2.1

# black==20.8b1
# dj-rest-auth==1.0.7
# django-allauth==0.42.0
# django-silk==4.0.1
# djangorestframework-camel-case==1.1.2
# factory_boy==2.12.0
# gunicorn==20.0.4
# pytest-cov==2.9.0
# pytest-django==3.9.0
# python-dateutil==2.8.1
# pytz==2020.1
# tox==3.15.1
# urllib3==1.26.5
